feat: extract sort to hook to generically use it
